fixes #3 fixes #15
A character can now hold an ItemStack
A hauling job now requires a Character to pick up an item from a pickup location and drop it off at another tile.
Harvesting a plant now creates a hauling job to two tiles south. This could later on be a container for food.